feat: renamed images to files
